+++ Capple/Capple/Data/Network/ApiEndpoints.swift
@@ -23,6 +23,9 @@ enum ApiEndpoints {
         
         // MARK: - 답변
         case answersOfQuestion = "/answers/question"
+        
+        // MARK: - 멤버
+        case signIn = "/members/sign-in"
     }
 }
