+++ Capple/Capple/Presentation/SignInView/ViewModel/AuthViewModel.swift
@@ -14,6 +14,8 @@ class AuthViewModel: ObservableObject {
     @Published var authorizationCode: String = ""
     @Published var name: String = ""
     
+    var signInResponse: MemberResponse.SignIn = .init(accessToken: nil, refreshToken: nil, isMember: false)
+    
     // Apple 로그인 요청 처리
     func appleLogin(request: ASAuthorizationAppleIDRequest) async {
         // Apple 로그인 요청을 처리하는 코드
@@ -28,18 +30,25 @@ class AuthViewModel: ObservableObject {
             print("Apple Login Successful")
             switch authResults.credential {
             case let appleIDCredential as ASAuthorizationAppleIDCredential:
-                let userIdentifier = appleIDCredential.user
-                let fullName = appleIDCredential.fullName
-                let name = (fullName?.familyName ?? "") + (fullName?.givenName ?? "")
-                let email = appleIDCredential.email
-                let identityToken = String(data: appleIDCredential.identityToken!, encoding: .utf8)
-                let authorizationCode = String(data: appleIDCredential.authorizationCode!, encoding: .utf8)
+                // let userIdentifier = appleIDCredential.user
+                // let fullName = appleIDCredential.fullName
+                // let name = (fullName?.familyName ?? "") + (fullName?.givenName ?? "")
+                // let email = appleIDCredential.email
+                // let identityToken = String(data: appleIDCredential.identityToken!, encoding: .utf8)
+                let authorizationCode = String(data: appleIDCredential.authorizationCode!, encoding: .utf8) ?? "인증 코드 생성 실패"
                 
                 DispatchQueue.main.async { /*[weak self] in*/
-                    print("Name: \(name)")
-                    print("Email: \(email)")
-                    print("IdentityToken: \(identityToken)")
+                    // print("Name: \(name)")
+                    // print("Email: \(email)")
+                    // print("IdentityToken: \(identityToken)")
                     print("AuthorizationCode: \(authorizationCode)")
+                }
+                
+                // 로그인 요청
+                Task {
+                    let signInResponse = try await NetworkManager.requestSignIn(request: .init(code: authorizationCode))
+                    self.signInResponse = signInResponse
+                    print(signInResponse)
                 }
                 
             default:
